+++ .php-cs-fixer.dist.php
@@ -20,6 +20,7 @@ use PhpCsFixerCustomFixers\Fixers;
 $finder = Finder::create()
     ->files()
     ->in([
+        __DIR__.'/api/src',
         __DIR__.'/examples',
         __DIR__.'/src',
         __DIR__.'/tests',